backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.3.6
certifi==2026.1.4
cffi==2.0.0
channels==4.2.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.13.4
cron_descriptor==2.0.6
cryptography==46.0.5
daphne==4.1.0
Django==5.2.10
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-prometheus==2.4.1
django-timezone-field==7.2.1
flower==2.0.1
gunicorn==23.0.0
humanize==4.15.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==26.0
paho-mqtt==2.1.0
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
Pygments==2.19.2
PyJWT==2.8.0
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-cov==6.1.0
pytest-django==4.10.0
pytest-mock==3.14.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-json-logger==4.0.0
pytz==2025.2
redis==7.0.1
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tornado==6.5.4
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
zope.interface==8.2
confluent-kafka==2.3.0